requirements/base.txt
albumentations>=1.1.0
antialiased-cnns
av>=10.0.0
einops>=0.3.2
freia>=0.2
imgaug==0.4.0
jsonargparse[signatures]>=4.3
kornia>=0.6.6,<0.6.10
matplotlib>=3.4.3
omegaconf>=2.1.1
opencv-python>=4.5.3.56
pandas>=1.1.0
pytorch-lightning>=1.7.0,<1.10.0
seaborn
timm>=0.5.4,<=0.6.12
torchmetrics==0.10.3
